promise: Refactor toCompletionStage into PromiseImpl
