+++ org.osgi.util.promise/src/org/osgi/util/promise/DeferredPromiseImpl.java
@@ -23,7 +23,6 @@ import static java.util.Objects.requireNonNull;
 import java.lang.reflect.InvocationTargetException;
 import java.util.NoSuchElementException;
 import java.util.concurrent.Callable;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ScheduledFuture;
@@ -760,45 +759,6 @@ final class DeferredPromiseImpl<T> extends PromiseImpl<T> {
 				v = null;
 			}
 			tryResolve(v, f);
-		}
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public CompletionStage<T> toCompletionStage() {
-		CompletableFuture<T> completableFuture = new CompletableFuture<>();
-		onResolve(new ToCompletionStage(completableFuture));
-		return completableFuture;
-	}
-
-	/**
-	 * A callback used for the {@link #toCompletionStage()} method.
-	 * 
-	 * @Immutable
-	 * @since 1.2
-	 */
-	private final class ToCompletionStage
-			implements Runnable, Result<T> {
-		private final CompletableFuture<T> completableFuture;
-
-		ToCompletionStage(CompletableFuture<T> completableFuture) {
-			this.completableFuture = requireNonNull(completableFuture);
-		}
-
-		@Override
-		public void run() {
-			result(this);
-		}
-
-		@Override
-		public void accept(T v, Throwable f) {
-			if (f == null) {
-				completableFuture.complete(v);
-			} else {
-				completableFuture.completeExceptionally(f);
-			}
 		}
 	}
 }
